await resolveConfigureWebpack whenever called
resolveConfigureWebpack is an async function that returns the promise of the
require statement. But all of the locations where resolveConfigureWebpack was
called, were not awaited. This meant that users of the plugin were not able to
install their own webpack configuration because the contents of configure-webpack.js
was never actually read and added to the vue-cli configuration object.

This commit adds the awaits for all of the locations where the function is called.
Unfortunately there are no unit tests that test these functions and I didn't know
how to go about adding unit tests for this area. So there are no new tests for
this change.

However, I believe it is important to get this out to users of the plugin quickly
because without this change the plugin is virtually useless for any substantial vue
app.